requirements.txt
pygame-ce>=2.5.1
mili-ui>=1.0.4
moviepy>=2.1.2
# OPTIONAL
pypresence>=4.3.0
pywebview>=5.4
youtube-search-python>=1.6.6
